Improve check that user has run setup script
The run-escape-variants.sh script contains a check that the
genome file has been setup using setup-escape-variants.sh.
This check assumed that `bwa idx` creates a file with a ".fai"
extension. This doesn't always seem to be the case for `bwa`.

Changes here simplify this check to just look for a ".dict" file
that is explictly created in setup.

Fixes #29