Adds pangolin step
Adds a step to the end of the escape variants pipeline that runs pangolin on consensus fasta files(*.cleaned.fasta).
This includes a new requirement of pangolin version 2.3.2.
Adds optional -p project name argument to run-escape-variants.sh. The project name field defaults to sars-cov2. This controls the name used in the pangolin output files. The run-dds-escape-variants.sh uses the input project name for the -p argument when calling run-escape-variants.sh.

The pangolin step creates two CSV output files. Empty pangolin CSV files are not copied to the output directory and uploaded to DDS. A message will be included in the log when a pangolin CSV file is empty.

Fixes #34