Add Snakefile to setup genome
Adds a Snakemake workflow to run setup steps on the genome.
The genome is named NC_045512.fasta and stored in a resources
directory per Snakemake docs recommendation:
https://snakemake.readthedocs.io/en/stable/snakefiles/deployment.html
The Snakefile consists of an "all" rule that specifies the
files to generate. These are the bwa index files, samtools index
file and the picard dictionary file. There are three
rules for creating the various files.
Instead of a single environment each rule has it's own conda
environment. Snakemake handles creating and using these
environments when snakemake is run with the --use-conda flag.
The rules specify a location for their log files.
Note the shell "&>{log}" part that saves command output to the
log files.

The docs also recommend putting each rule in a separate file
but for clarity (while we are getting started) I left the rules
in the main Snakefile.

To run on a slurm cluster requires a "slurm" profile.
I will address this need in a later change.
Right now I have installed snakemake on my laptop and run
the pipeline like so:
```
snakemake --cores 1 --use-conda
```