Fixes sam file not found error
The problem was extracting the sample name from the file.
The sample name is created by taking the basename of the FASTQ
filename then returning the part before the first "_".

Changes here create a dictionary in workflow/rules/common.smk
that returns a FASTQ filename for a sample name. This change and
using the sample name for the trimmed read filename removed the
need for the readsuffix config option. The remove_adapters rule
now specifies it's output and the remove-nextera-adapters.sh
moves it's output file to the appropriate name.

The output bam file from the create_bam_from_sam is now stored
in it's own directory. This was done after receiving a a warning
from snakemake to do this to prevent ambiguity.

Fixes #69